requirements.txt
tensorflow==2.3.0
keras==2.4.3
scikit-learn
pickle-mixin
pillow
opencv-contrib-python-headless
mtcnn
numpy==1.16.0
